Show Map Filters - se paso a properties el valor de radius para levantar los activeSessions. - se paso a double en vez de float los campos x,y de MapReducePoint. - se actualiza un poco el archivo TODO de docs. - se pone el logging en INFO en vez de DEBUG.

+++ masterProject/core/src/main/java/com/uade/pfi/core/repositories/mapreduce/MapReducePoint.java
@@ -1,11 +1,11 @@
 package com.uade.pfi.core.repositories.mapreduce;
 
 public class MapReducePoint{
-	float x,y;
-	public void setX(float x) {
+	double x,y;
+	public void setX(double x) {
 		this.x = x;
 	}
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 }
